requirements.txt
numpy
pandas
scikit-learn
flask
pillow
keras
numpy
tensorflow